Cleaning up tests related to Costs
